@@ -1,64 +1,76 @@
 {
-  "name": "portafolio",
-  "version": "3.1.0",
+  "name": "rojasarmando.com",
+  "version": "4.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "test": "ng test",
-    "lint": "eslint '*/**/*.{js,ts,tsx,html}' --quiet --fix",
-    "e2e": "ng e2e"
+    "watch": "ng build --watch --configuration development",
+    "lint": "eslint src",
+    "lint:fix": "eslint --fix src",
+    "test": "jest",
+    "test:coverage": "jest --coverage"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.0.0",
-    "@angular/common": "^12.0.0",
-    "@angular/compiler": "^12.0.0",
-    "@angular/core": "^12.0.0",
-    "@angular/forms": "^12.0.0",
-    "@angular/platform-browser": "^12.0.0",
-    "@angular/platform-browser-dynamic": "^12.0.0",
-    "@angular/router": "^12.0.0",
-    "@ngx-translate/core": "^13.0.0",
-    "@ngx-translate/http-loader": "^6.0.0",
+    "@angular/animations": "^17.0.0",
+    "@angular/common": "^17.0.0",
+    "@angular/compiler": "^17.0.0",
+    "@angular/core": "^17.0.0",
+    "@angular/forms": "^17.0.0",
+    "@angular/platform-browser": "^17.0.0",
+    "@angular/platform-browser-dynamic": "^17.0.0",
+    "@angular/router": "^17.0.0",
+    "@ngx-translate/core": "^15.0.0",
+    "@ngx-translate/http-loader": "^8.0.0",
     "animate.css": "^4.1.1",
-    "bootstrap": "^5.1.2",
-    "jquery": "^3.6.0",
-    "ng-particles": "^2.18.1",
-    "ngx-spinner": "^12.0.0",
-    "ngx-typed-js": "^2.1.0",
+    "bootstrap": "^5.3.2",
+    "jquery": "^3.7.1",
+    "ng-particles": "^3.12.0",
+    "ngx-spinner": "^16.0.2",
+    "ngx-translate": "^0.0.1-security",
+    "ngx-typed-js": "^2.1.1",
     "ngx-wow": "^2.0.1",
-    "rxjs": "~6.6.0",
-    "sweetalert2": "^11.1.7",
-    "tslib": "^2.0.0",
-    "tsparticles": "^1.35.1",
+    "rxjs": "~7.8.0",
+    "sweetalert2": "^11.9.0",
+    "tslib": "^2.3.0",
+    "tsparticles": "^2.12.0",
     "wowjs": "^1.1.3",
-    "zone.js": "~0.11.3"
+    "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^12.1.1",
-    "@angular/cli": "^12.2.7",
-    "@angular/compiler-cli": "^12.0.0",
-    "@types/jasmine": "~3.6.0",
-    "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "^4.27.0",
-    "codelyzer": "^6.0.0",
-    "eslint": "^7.29.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-html": "^6.2.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "^6.3.4",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.0.3",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "prettier": "^2.3.2",
-    "prettier-eslint": "^12.0.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.3.2"
+    "@angular-devkit/build-angular": "^17.0.0",
+    "@angular/cli": "^17.0.0",
+    "@angular/compiler-cli": "^17.0.0",
+    "@types/jest": "^29.5.8",
+    "eslint": "^8.53.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-plugin-import": "^2.29.0",
+    "jest": "^29.7.0",
+    "jest-preset-angular": "^13.1.4",
+    "karma-coverage": "^2.2.1",
+    "prettier": "^3.0.3",
+    "typescript": "~5.2.2",
+    "typescript-tslint-plugin": "^1.0.2"
+  },
+  "jest": {
+    "preset": "jest-preset-angular",
+    "roots": [
+      "<rootDir>/src"
+    ],
+    "setupFilesAfterEnv": [
+      "<rootDir>/setupJest.ts"
+    ],
+    "testPathIgnorePatterns": [
+      "<rootDir>/node_modules/",
+      "<rootDir>/dist/",
+      "<rootDir>/src/test.ts"
+    ],
+    "globals": {
+      "ts-jest": {
+        "tsConfig": "<rootDir>/tsconfig.spec.json",
+        "stringifyContentPathRegex": "\\.html$"
+      }
+    }
   }
 }
