@@ -9,78 +9,19 @@ export class CVComponent implements OnInit {
   public animateBtn: Boolean = false;
   public id = 'tsparticles';
   public particlesOptions = {
-    background: {},
     fpsLimit: 60,
-    interactivity: {
-      detectsOn: 'canvas',
-      events: {
-        onClick: {
-          enable: true,
-          mode: 'push'
-        },
-        onHover: {
-          enable: true,
-          mode: 'repulse'
-        },
-        resize: true
-      },
-      modes: {
-        bubble: {
-          distance: 100,
-          duration: 2,
-          opacity: 0.5,
-          size: 40
-        },
-        push: {
-          quantity: 4
-        },
-        repulse: {
-          distance: 60,
-          duration: 0.3
-        }
-      }
-    },
     particles: {
       color: {
-        value: '#ffffff'
+        value: "#000"
       },
       links: {
-        color: '#ffffff',
-        distance: 150,
         enable: true,
-        opacity: 0.5,
-        width: 1
-      },
-      collisions: {
-        enable: true
+        color: "#000"
       },
       move: {
-        direction: 'none',
-        enable: true,
-        outMode: 'bounce',
-        random: false,
-        speed: 2,
-        straight: false
-      },
-      number: {
-        density: {
-          enable: true,
-          value_area: 100
-        },
-        value: 10
-      },
-      opacity: {
-        value: 0.1
-      },
-      shape: {
-        type: 'circle'
-      },
-      size: {
-        random: true,
-        value: 4
+        enable: true
       }
-    },
-    detectRetina: true
+    }
   };
 
   constructor() {}
