@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { LangService } from 'src/app/services/lang.service';
+import { LangService } from '../../services/lang.service';
 @Component({
   selector: 'app-header',
   templateUrl: './header.component.html',
