@@ -1,5 +1,5 @@
 export interface SkillInterface {
   img: String;
-  title: String;
-  description?: String;
+  title: string;
+  description?: string;
 }
